changes to perceptual.
Overloaded the get resolution function to allow a bool to be passed in.
use the out keyword when passing the bool to make the bool get a success
return value.
True if the resolution is legitimate and false if it isn't.

Added gesture tracking for the circle hand motion. Added a corresponding
circled() function
to poll if the gesture has taken place.

The Volume is now saved every time the volume is changed. Use getVolume
to know the current volume.
The volume is initialized to 0.3f

Altered getOptionVoicedAsString to loop through an arbitrarily large
list of options. now all one has
to do is add more strings to the commands string array and you receive
more strings than before.

found voiceheard bool in PerCVoice was redundant, so it's been removed.